Split fuelup_cli into it's own file
