+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod commands;
 pub mod constants;
 pub mod download;
 pub mod file;
+pub mod fuelup_cli;
 pub mod ops;
 pub mod path;
 pub mod settings;
